pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gsd-bridge"
version = "3.0.0"
description = "Bridge between GSD plans and Codex execution state"
readme = "codex-bridge-template.md"
requires-python = ">=3.9"
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "ruff>=0.4",
]

[project.scripts]
gsd-bridge = "gsd_bridge.cli:main"

[tool.setuptools.packages.find]
include = ["gsd_bridge*"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_unused_configs = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py39"
